Fixed FAQ accordion
The FAQ accordion was only working if the domainname and master password fields were filled in because of setfocus().  That means a click anywhere else, such as one of the FAQs moved the focus to one of those fields.

The problem was using onblur, which fires whenever the focus leaves the field for any reason.  I changed that to onmouseleave, which solved the problem.

I also commented out the FAQ answer on personas, since that's hidden until I get a chance to test it.