Put bookmark settings into hpSPG
The idea is that I'll use the bookmark to get my settings for a given domain name and then use localStorage unless something changes.  In order to do that, I need to make sure any changes from the extension are included in what goes to localStoage when the user clicks Remember.