Updated license agreement
I was working for HP when I did the first version of this code, and they required a specific license agreement be included in each file.  I was unable to reach the legal department at either HP Inc. or HP Enterprise to find out how to change it.  This is my attempt.