Update instructions
The web page doesn't have access to bookmarks, so the only site data it has is what's been stored in localStorage.  The original label did not make that clear.

That change meant that the handling of download is different for the web page and the extension.

Also, the antiphishing description belongs in the Common section, not the Overview.