Extract domain name from pasted URL
People should paste the domain name instead of manually typing it to make phishing attacks harder.  However, the metadata is indexed by domain name.  I fix that by taking the URL in the domain name field and extracting the domain name from it.  Notice the setTimeout, which is needed because onpaste fires before the value is set.