Change handling of paste bookmark
I was unhappy with what happened when you pasted a bookmark after pasting the login page URL.  The bookmark field cleared, and the site and user name fields got populate so fast you didn't know anything happened.  I'm not completely happy with this version, but it's clear that something changed.

I also wan't clearing the site password and site and user name fields when the user pastes a new login page URL.