Change sitename working
The problem was putting the value from the form into SitePassword.settings for numeric fields.  element.value is a string, but the cached data has integers for these fields.  Hence, checking for changes always succeeds even when it shouldn't.  The fix is to convert those strings to integers before putting them into SitePassword.settings.