First attempt to use favicon to warn pw on clipboard
When the user clicks the clipboard icon on the site password field, I want to change the favicon to alert the user that the password is on the clipboard.  I can't change it back automatically, but the user can clear the warning by clicking anywhere on the SitePassword page.

The problem is that I can't get the favicon to change from JavaScript, even though I can from the console.  Right now I'm stuck.