Reduce frequency of phishing false positives
One of the real annoyances of SitePassword is the phishing warning when a site has multiple login pages with different domain names.  That wouldn't be a problem if you could always figure out the owner from the URL.

It is virtually impossible to do that  There is just too much variation in where it sits in the domain name.  For example, amazon.com vs amazon.co.uk.  The best you can do is a 90% solution.

That's what I did here.  I cribbed a line of JavaScript from StackOverflow that will get it right most of the time.  Increasing the number of TLDs in the list will get it right for even more sites, but the existing list is a really good start.