New logo
After a long back and forth with Doug, I found a logo that he likes.  This may not be the final version, though.  Here, the favicon shares the background color of the SitePassword form, and the version that appears on the form has a white background.  I like that better than the colored background, but I'll wait to hear what Doug has to say.