Handle any protocol and domain name in bookmarks
Since Safari doesn't support the download API, I'm opening a new window to show the settings table that I would have downloaded.  Unfortunately, Chrome decided to disallow opening a top level window.  Ok, then I'll open the table in an iframe.  Oh no, I can't control the protocol and domain of anything in an <a href=>.  It must be that of the extension.

My solution is to parse the URL string to extract the settings part.  It's complicated by the fact that bookmarks start with ssp:// and no domain name while the URLs in the iframe are of the form chrome://extensionid/.  This version appears to get them goth right.