Add datalist to site and user name fields
The latest version of the extension has datalists for the site and user name fields.  For the site name field, the datalist serves as a reminder of previously chosen site names.  For the username field, it makes it easier to enter a previously used username.  This commit adds that feature to the web page.