Add phishing warning when clicking bookmark
One problem with clickable bookmarks is that you lose a potential phishing warning because you don't paste the login page URL.  I could force you to do that, but that's inconvenient on a mobile device when you're getting a password for an app.  Instead, I made pasting the URL optional by adding a warning when the form is filled in because you clicked a bookmark.