Use innerText instead of innerHTML
The latter is dangerous if you're including data from an untrusted source.  In this case, I'm including the domain name of a potential phishing site.  Domain names should be OK even without sanitizing, but innerText is guaranteed to be safe.

Getting the formatting right isn't easy with innerText, but I was able to get what I want by breaking the warning message into parts.