Fixed a bug handling special characters
The extension needs to handle special characters carefully because they go through a URL encode step.  The web page doesn't have that problem, but it does need to handle bookmarks created by the extension.  As a result, the web page needs to do the same careful handling of special characters.  I hadn't done that, which resulted in the web page generating different passwords even when not using special characters.