Start of test framework
I'm tired of spending an hour manually testing changes, so I'm automating them.  One important part is the ability to continue a test following a reload of the page.  I'm using localStorage to tell me if I should run the first part of the test or a subsequent one following a page reload.