Test forget
While testing something else, I found a bug in forgetting a domain name.  The problem is that it leaves the site name even if no domain names refer to it.

The bug shows up when you enter a previously used site name that isn't associated with any domain names and the corresponding user name appears.  I can't make that happen in the test code, so I just look at the database to see if there's an entry for a site name that shouldn't be there.