Eliminate help for code field
The help text for the input fields on the main panel appears to the right of the panel, which currently pushes the settings panel to the right.  The plan is to avoid that pushing by closing the setting when you open a help item, but that won't work for the code field, which is on the settings panel.  The easiest solution to this dilemma is to remove the 3 dots menu from the code field.