Automatically move elements
Moved settings and instructions out of main panel so that they now move from next to the main panel to under when there isn't enough horizontal space.

I also took the border off the main panel because it grew to make room for the instructions when they were next to the settings but didn't shrink when the instructions closed.