Got PBKDF2 working
with a bit of help from Copilot.  (See the return statements on lines 134 and 137.)

This version does 25,000 hash iterations and up to 1,000 more to try to find an acceptable password.  Computing a password takes about 100 ms, which is comfortably fast enough to update the site password on each keystroke.  A failed attempt to compute a password takes about 0.5 seconds.  That's long, but not uncomfortably so.  It's a tradeoff because I can sometimes find a valid password with thousands of extra iteration, but that can take several seconds.