Hardened super password
by increasing the iteration count of PBKDF2 so that the algorithm typically runs for 1-2 seconds.  That means I no longer compute on each key stroke, only when you mouse out of the field.  I also produce a 2K string so that there are lots of super password guesses that produce a given site password.