Improved hardened super password
The previous commit for hardening the super password created a very long key with PBKDF2 but then did a SHA-256 hash until it found a valid password for the site.  The problem is that approach loses the benefit of the long key.

This version doesn't hash to find a valid site password.  Instead, it rotates the character set one character and tries again if that doesn't produce an acceptable password, repeating until it does or has rotated the full 64 times.

This approach works with high probability, but it's not as successful as the original algorithm.  Fortunately, the difference only shows up in extreme cases, such as requiring a lot of special characters.  While the old algorithm can find a valid 16-character site password with 6 special characters, the new one can't always find one with more than 4.  Fortunately, no sites have requirements like that.