Two changes to algorithm
The previous commit rotated the character set if the computed password wasn't acceptable.  It failed to find a valid password more often than the repeated hashing did.

This commit creates a string of about 1,000 characters (limited by the latency requirement) from the 64Mb hash and picks the first try from the head of that string.  If it doesn't pass, it strips the first candidate character and tries again.  It takes about 0.5 seconds to discover there is no valid password.

The second change is to modify the character set construction to increase the chance that a number or special character will be selected.  To do that I repeat the integers once and the special characters until the string is at least 20 characters.  I had to change the bits to conversion routine be Base128 instead of B64, but I didn't change the function name.