Change hashing parameters
I realized that taking a small subset of the characters of a very long hash doesn't help protect your super password.  There are lots of collisions, but very few of them are reasonable super passwords.  That's why I'm changing the hash parameters to produce a shorter hash with more iterations.

I picked the length of the hash to be long enough that I can usually find a legal password even for extreme settings, e.g., 8 special characters, and still meet the 200 ms latency requirement.