Fix conversion from bytes to chars
The hash function produces a byte array that was originally converted to characters using binl2b64() in sha256.js.  I now increase the ability to generate passwords that require more special characters or digits by putting more of them in the character array.  As a result the conversion needs to be done differently.  I make that conversion particularly simple by padding the array to 256 characters.  Now each byte of the array is an index into the character array.  No bit fiddling needed.