A few fixes
Fixed an async error in bg.js where a pwfail message would flash on the window.  The problem was when a keyboard action triggered 2 events.  For example, superpw.onblur and allowspecialscheckbox.oncheck.  Due to async settings.allowspecial was inconsistent with the character set.  The solution was to move generating the character set into the closure.

I also changed the hardening when looking for a valid password from doubling the exponent of h to adding 2 to it.

I also removed several redundant calls to generate password in ssp.js.