Limit site passwords to 40 characters
One feature of SitePassword is that it generates lots of pre-images, super passwords that produce the same site password for one site but not for any others.  Each pre-image must be tested online, discouraging the attack.

An attack that weakens that protection is for the adversary's site to require a very long site password.  A 40-character site password is strong enough for all legitimate sites but short enough not to weaken the pre-image protection.