Change hash parameters
I was doing a lot of iterations to create a 2Kb key.  However, there is apparently nothing wrong with doing 1 iteration to create a 64Mb key.  The latter produces more pre-images in the same amount of time.  I can fail to find a password fast enough for users if I convert  bytes to 256 characters.