Change search for valid password
I was converting 2Kb of the key to characters and using a sliding window to find a valid password.  This version hashes until one is found or too much time has passed.

The first hash is intentionally slow to increase an adversary's work factor.  If the generated password doesn't pass, then I hash again with a fast hash function, one iteration of PBKDF2 with a key size of 2Kb.  I use a fast hash function here to get as many tries as possible in the allotted time.