Increased the key size when iterating
I want a fast hash function when iterating to find a valid password so I get as many tries as possible in the available time.  However, the key size has to be big enough to generate a lot of pre-images.  The settings in this commit result in about 1 ms per try while having a lot more pre-images than before.