Diacono algorthim for bytes to chars
James proposed an algorithm that always produces a valid site password using bytes selected from the hash.  It chooses one byte to determine which character in the site password it's selecting and the next byte to select which character goes there.

Note that this approach may allow the adversary to filter pre-images offline.  See Collisions.pages for details.