Switch from byte/char to uint16/char
If I use a byte/char in the conversion from bits to characters, there are 4 bytes that produce the same character with a 64 element alphabet.  In other words, it gives 3 pre-images per character in the site password.  If I use uint16s, there are a lot more pre-images.  Since the hash is 64 Mb, I won't run out of bits in the allotted time..