Use only bytes to index into character array
I originally thought that I could generate more collisions if I used more bits when indexing into the character array, but that's not the case.  The number of collisions only depends on the alphabet size.  That being the case, there's no point in using anything other than a byte as an index, since larger indexes make subsequent iterations take longer.