Add entropy estimate to superpw strength meter
A long super password is better than a shorter one unless the longer one is too easy to guess.  The strength meter gives some guidance, but more information is better.  This commit uses the number of guesses given by zxcvbn to tell the user the number of random characters the chosen site password is equivalent to.