Stop iterating based on count, not time
I've given myself about 150 ms to decide that I can't find an acceptable password, so it seems reasonable to use that time as the condition on the loop.  Unfortunately, the JavaScript garbage collector sometimes causes a false negative.

I'm just going to use a fixed number of iterations without even tuning the number to the speed of the machine.  The theory, which is only partly borne out by experience, is that you won't ever find one if you don't find it in 150 tries.