Fix a big boo boo
I was attempting to make PBKDF2 memory intensive by computing a very big key.  That succeeded for the library code, but an adversary can produce a version that is not memory intensive but returns the same value.

There was a chance that I could preserve the memory intensive feature because the common use stores the key.  However, my attempts failed.  For example, I could take the last 32 bits of the key to decide where to start taking bytes for the password, but then the adversary could just run the algorithm twice, once to find the last 32 bits and once to compute the password.  In neither run would the whole key need to appear in memory.