Directly construct a valid password
Rather than iterating the hash to find a valid password, I can construct one directly, but it's a bit tricky.  I can insert any required characters, e.g., minimum number of digits, into the string and fill in the rest with characters selected from the full alphabet, but then the site passwords have a known pattern.

The solution to that problem is to permute the elements using a deterministic random shuffle.  I'm using the same bytes of the key that I used to select the characters.

Having shuffled the characters, they may not meet a "starts with a letter" condition.  I fix that problem by shuffling all characters but the first.