Modify strength meter
to encourage people to use a super password that is less guessable than their site passwords.  I use zxcvbn report.guesses_log10 as a proxy for the entropy and give the site password a boost.  The hope is that people will choose a super password that is effectively 4 characters longer than the site password.  The result will be to force someone attempting to guess the super password make so many online tests to verify it's the right one that guessing becomes impractical.