Use different strength calculations
The super password is picked by a user, so the zxcvbn strength calculation is appropriate.  The site password is calculated from a good hash function, so the zxcvbn algorithm grossly underestimates its strength.

The risk of using the same algorithm for both is that the user will pick much stronger site passwords than necessary, making offline guessing of the super password practical.