Adjusted strength meter
The strength meter serves 2 purposes.  One is obvious, to warn the user of a weak super or site password.  The second purpose is hidden from the user, to encourage the use of a super password with more entropy than the shortest site password that is rated to take longer than 1,000 years to guess, which is 12.

The value chosen on line 235 means that for a super password rated to take over 1,000 years to guess there will be 16K guesses that will produce the same site password for a single site.  An adversary will have to try each such guess online, which makes the guessing attack impractical.