Adjust superpw strength penalty
This commit includes code to calculate what fraction of random strings have full entropy.  It's a surprisingly small value, 92% for 12-character strings, 87% for 20-character strings.  As a result, the superpw strength penalty doesn't need to be nearly as large as I feared because an adversary would have to guess a larger subset of all possible super passwords.