Move "await Promise.resolve()" and add rel attribute
It's used enforce ordering when one branch of a test has an await and another one doesn't. I had it at the top of the async function, but I think it belongs on the branch(es) without an await.

I should have had a separate commit for the rel attribute changes. They prevent cross-site scripting when you have an href. I don't think it's necessary for data URLs, but adding it does no harm.