Fix special characters repeat
There are often fewer special characters allowed than there are digits.  As a result, site passwords that require a lot of special characters end up using the weaker deterministic algorithm.  I was repeating the special characters when constructing the alphabet, but that weakens site passwords that don't need the deterministic algorithm in analogy to modulo bias.

The only time the deterministic algorithm gets invoked is when the set of special characters is small and a lot of them are required, conditions I've never seen in the wild.   As a result, I've decided not to repeat the special characters.