Protect super password
When you edit your super password, SitePassword recursively hashes it for 15 or so seconds.  The result is stored in session storage and used as part of the salt for calculating your site password.  An adversary would have to repeat that long calculation for every guess.  Even if the adversary could get the cached value, you are still protected by the strength of your super password and the slow hash used for the base calculation.