Change protect UX
Both Doug and Dale thought 10-15 seconds was too long.  Dale pointed out that 5 seconds is the upper limit for interactive things that are usually fast, so I picked parameters so the delay is about 2 seconds on my old laptop.

Neither of them liked the warning box idea for telling the user what's going on.  Instead, I'm using Dale's idea of using the strength meter as a progress bar.  It's approximate because I pick how long the meter changes, but I can only specify the number of iterations.  I also add a label in red "Protecting" the way I put "copied!" on when the clipboard icon is clicked.

I also don't want a site password showing up until the cache value has been computed, such as when you are typing a super password.  I put labels in that field to tell you what to do.