Changed some parameters
I was caching a 32 character value, which has more entropy than any likely super password.  That means that a guess for the super password that produces this value is the right super password..

I'm now caching 12 characters, so many guesses for the super password will give that value.  The adversary will need a different test to know if a hit is the right value.