Use progress bar
while computing the cached value.  I'm using it with no value specified, which makes the colored part swing left and right rather than progress forward.  The issue is knowing how fast to adjust the value.  Too slow, and the bar hardly moves; too fast, and the bar reaches 100% before the computation is done.