Add a new option
You might be using the web page to get a password when using a friend's machine.  In that case, you probably don't want to leave you super password lying around.  This commit adds a new option that lets you decide the behavior you want.  It defaults the safe behavior with the option checked.