Fix bug in phishing warning with bookmarks
The form wasn't being filled when selecting "Same account" if the bookmark domain name didn't match the form domain name because of a difference from the extension.  There the only source of settings is the database, but here it can also come from the bookmark.  Also, I don't want to implicitly save the settings  on the web page like I do in the extension.