Move warnings
They were to the left of the main form, but that makes things jump to the right when a warning appears.  I tried putting the warnings to the right of settings, but they won't appear if the window is too narrow.  As a compromise, I'm putting the warnings between the main form and the settings.  If you don't notice the horizontal scroll bar, you'll thing that the warning replaced the settings panel.  I think that's OK because of the Close button.