Fix handling of pasted domain name
When you open the page by clicking on a bookmark, you're given the opportunity to paste a URL to check for phishing.

I had 2 problems.  First, I cleared the domain name field if you pasted an incorrectly formatted URL.  That led to phishing warning if you then pasted a properly formatted URL that matched the bookmark's domain name.  Second, I allowed typing in that field.