Allow bookmark paste without domain name
In order to get a phishing warning, you had to paste the URL of the login page before pasting a bookmark.  That requirement makes it harder to use a bookmark because you have to remember to copy the URL before you copy the bookmark.  This commit allows you to paste the bookmark without first providing a URL.  However, if you do provide a URL after pasting the bookmark, the form will be cleared.